requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-active-link==0.1.8
django-ckeditor==6.5.1
django-js-asset==2.0.0
idna==3.4
Pillow==9.5.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.3
